Prepare ros control
