+++ webots_ros2_control/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.5)
-project(webots_ros2_bridge)
+project(webots_ros2_control)
 
 # Default to C99
 if(NOT CMAKE_C_STANDARD)
@@ -20,53 +20,75 @@ find_package(ament_cmake REQUIRED)
 find_package(hardware_interface REQUIRED)
 find_package(pluginlib REQUIRED)
 find_package(rclcpp REQUIRED)
+find_package(webots_ros2_cpp REQUIRED)
 
+# Webots
+include_directories(
+  include
+  $ENV{WEBOTS_HOME}/include/controller/c
+  $ENV{WEBOTS_HOME}/include/controller/cpp
+)
 
-# ros2_control
-# ros2_control: compile
+# webots_ros2 Plugin
 add_library(
-  ${PROJECT_NAME}_control
+  ${PROJECT_NAME}
   SHARED
   src/Ros2Control.cpp
 )
 target_include_directories(
-  ${PROJECT_NAME}_control
+  ${PROJECT_NAME}
   PRIVATE
   include
 )
 ament_target_dependencies(
-  ${PROJECT_NAME}_control
+  ${PROJECT_NAME}
   hardware_interface
   pluginlib
   rclcpp
+  webots_ros2_cpp
 )
+pluginlib_export_plugin_description_file(${PROJECT_NAME} webots_ros2_control.xml)
 
-pluginlib_export_plugin_description_file(hardware_interface wb_ros2_control.xml)
-
-# ros2_control: install
-install(
-  TARGETS ${PROJECT_NAME}_control
-  DESTINATION lib
+# ros2_control System
+add_library(
+  ${PROJECT_NAME}_system
+  SHARED
+  src/Ros2ControlSystem.cpp
 )
-install(
-  DIRECTORY include/
-  DESTINATION include
+target_include_directories(
+  ${PROJECT_NAME}_system
+  PRIVATE
+  include
+)
+ament_target_dependencies(
+  ${PROJECT_NAME}_system
+  hardware_interface
+  pluginlib
+  rclcpp
+  webots_ros2_cpp
+)
+pluginlib_export_plugin_description_file(${PROJECT_NAME} webots_ros2_control_system.xml)
+
+# Common
+install(TARGETS
+  ${PROJECT_NAME}
+  ${PROJECT_NAME}_system
+  ARCHIVE DESTINATION lib
+  LIBRARY DESTINATION lib
+  RUNTIME DESTINATION bin
+)
+ament_export_include_directories(
+  include
 )
 
 if(BUILD_TESTING)
   find_package(ament_cmake_gtest REQUIRED)
 endif()
 
-# ros2_control: export
-ament_export_include_directories(
-  include
-)
-ament_export_libraries(
-  ${PROJECT_NAME}_control
-)
-ament_export_dependencies(
-  hardware_interface
-  pluginlib
-  rclcpp
-)
+#ament_export_dependencies(
+#  hardware_interface
+#  pluginlib
+#  rclcpp
+#)
+
 ament_package()
